finish cleaning up Swordfish.Library

+++ Swordfish.Library/Extensions/ColorExtensions.cs
@@ -1,5 +1,6 @@
 using System.Drawing;
 using System.Numerics;
+// ReSharper disable UnusedMember.Global
 
 namespace Swordfish.Library.Extensions;
 
@@ -8,19 +9,19 @@ public static class ColorExtensions
     public static Vector4 ToVector4(this Color color)
     {
         return new Vector4(
-            color.R / byte.MaxValue,
-            color.G / byte.MaxValue,
-            color.B / byte.MaxValue,
-            color.A / byte.MaxValue
+            (float)color.R / byte.MaxValue,
+            (float)color.G / byte.MaxValue,
+            (float)color.B / byte.MaxValue,
+            (float)color.A / byte.MaxValue
         );
     }
 
     public static Vector3 ToVector3(this Color color)
     {
         return new Vector3(
-            color.R / byte.MaxValue,
-            color.G / byte.MaxValue,
-            color.B / byte.MaxValue
+            (float)color.R / byte.MaxValue,
+            (float)color.G / byte.MaxValue,
+            (float)color.B / byte.MaxValue
         );
     }
 }
